+++ src/search/main.cpp
@@ -9,7 +9,7 @@ int main(int argc, char** argv) {
     State currentState = State();
 
     while(true) {
-        int timeToUse = 3;
+        double timeToUse = 3;
         MCTS search = MCTS(currentState, timeToUse);
         State bestMove = search.getBestMove();
         std::cout << bestMove.getStringBoard() << std::flush;
